Protocols: add `electronic_type` to the `get_builder_from_protocol`
This argument allows a caller to specify the electronic type of the
system, whether it is to be considered as a metal or insulator.
Additionally, the enum `ElectronicType` provides the value `AUTOMATIC`
which instructs the protocol that the type is unknown and the workflow
should attempt to automatically assess it. For now, only `INSULATOR` and
`METAL` are supported.