Protocols: implement protocol input generators for pw workflows
The `PwBaseWorkChain`, `PwRelaxWorkChain` and `PwBandsWorkChain`
implement the `get_builder_from_protocol` classmethod, that works just
as the `get_builder` method as in that it return a `ProcessBuilder`
instance for the relevant work chain. The major difference is that for
the new method, the builder is fully defined with all important inputs
set according to the specified protocol.

All three workflows come with three different protocols: `moderate`,
`precise` and `fast`, with `moderate` being the default which should
give a reasonably accurate result at moderate computational cost. The
`precise` protocol will give a more precise result but is more costly,
whereas the `fast` protocol can be used for testing but its results are
not to be blindly trusted as correct.

Note that only the values of the `moderate` protocol have been tested
but the other two have merely been extrapolated from there. They will
most likely need to be tweaked in the future after extensive testing.