Protocols: add `spin_type` to the `get_builder_from_protocol`
For now, only `SpinType.NONE` and `SpinType.COLLINEAR` are supported
even though non-collinear and spin-orbit are in principle also supported
by Quantum ESPRESSO itself. In the collinear case, the starting
magnetization will be determined automatically, unless it is overridden
by the caller through the `initial_magnetization` argument. It takes a
single starting magnetization for each kind present in the structure.
There is no automatic process implemented yet to automatically update
the structure to create as many kinds needed for the given initial
magnetization.